Rename places::storage::bookmarks::public to places::storage::bookmarks::public_node
