+++ components/places/src/storage/bookmarks/public_node.rs
